requirements/base.txt
boto3==1.9.144
Django==1.11.20
django-ordered-model==2.1.0
django-storages==1.7.1
Markdown==2.6.11
nltk==3.2.5
requests==2.21.0
scikit-learn==0.19.1
scipy>=0.14.0
xhtml2pdf==0.2.3
-r _base.txt
